requirements.txt
apiai==1.2.3
cachetools==2.0.0
catkin==0.7.6
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
gencpp==0.5.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.8
gennodejs==2.0.1
genpy==0.6.5
google-assistant-library==0.0.2
google-auth==1.0.2
google-auth-oauthlib==0.1.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
message-filters==1.12.7
numpy==1.13.1
oauthlib==2.0.2
olefile==0.44
Pillow==2.9.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
requests==2.18.3
requests-oauthlib==0.8.0
rosbag==1.12.7
rosboost-cfg==1.13.5
rosclean==1.13.5
roscreate==1.13.5
rosgraph==1.12.7
roslaunch==1.12.7
roslib==1.13.5
roslz4==1.12.7
rosmake==1.13.5
rosmaster==1.12.7
rosmsg==1.12.7
rosnode==1.12.7
rosparam==1.12.7
rospy==1.12.7
rosservice==1.12.7
rostest==1.12.7
rostopic==1.12.7
rosunit==1.13.5
roswtf==1.12.7
RPi.GPIO==0.6.3
rsa==3.4.2
RTIMULib==7.2.1
sense-hat==2.2.0
six==1.10.0
topic-tools==1.12.7
urllib3==1.22
Werkzeug==0.12.2
